// src/chains/moonbeam/types/generated/v1300.ts
export type CurrencyId = CurrencyId_SelfReserve | CurrencyId_OtherReserve

export interface CurrencyId_SelfReserve {
    __kind: 'SelfReserve'
}

export interface CurrencyId_OtherReserve {
    __kind: 'OtherReserve'
    value: bigint
}